feat: Add edit address functionality
This commit adds the necessary code changes to implement the edit address functionality. It includes the creation of a new HTML template for editing the address, as well as the corresponding view and URL mapping. The form fields are rendered dynamically using a loop, and error messages are displayed if there are any validation errors. The changes also include updating the URLs to use the new view for editing the address.

Closes #123