feat: Add brand filter functionality to products page
This commit adds the brand filter functionality to the products page. It includes the creation of a new HTML template, 'brand_filter.html', which is included in the 'products.html' template. The 'brand_filter.html' template loads the 'brandFilter.js' script using the {% static %} template tag. This change improves the user experience by allowing users to filter products by brand.

Closes #123