feat: Add pagination to NotificationView
This commit adds pagination to the NotificationView in the home/views.py file. The paginate_by attribute is set to 10, allowing the notifications to be displayed in smaller, more manageable chunks. This improves the user experience by reducing the amount of content on the page and making it easier to navigate through the notifications.

Closes #123