feat: Update button styles in top_five_product_by_date.html and most_products.html templates
This commit updates the button styles in the top_five_product_by_date.html and most_products.html templates. In the top_five_product_by_date.html template, the buttons for the time periods '1_month' and 'this_year' are removed. In the most_products.html template, a new CSS class 'see-more-btn' is added to the 'a' element to style the 'See More' button with a background color defined by the 'fourth-color' variable.

Closes #123